requirements.txt
beautifulsoup4==4.13.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
emoji==2.14.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.8
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
nltk==3.9.1
numpy==2.2.2
nvidia-nccl-cu12==2.25.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
soupsieve==2.6
textblob==0.19.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
Werkzeug==3.1.3
wordcloud==1.9.4
xgboost==2.1.3
